+++ packages/banana-react/src/index.ts
@@ -16,6 +16,7 @@ import { Rating } from './rating';
 import { Select } from './select';
 import { SelectOption } from './select-option';
 import { Stepper } from './stepper';
+import { Tooltip } from './tooltip';
 
 export {
   Button,
@@ -36,4 +37,5 @@ export {
   Select,
   SelectOption,
   Stepper,
+  Tooltip,
 };