+++ packages/banana/src/dropdown/index.test.ts
@@ -1,5 +1,5 @@
 import { aTimeout, expect, fixture, html } from '@open-wc/testing';
-import { sendKeys, sendMouse } from '@web/test-runner-commands';
+import { sendKeys } from '@web/test-runner-commands';
 import BDropdown from '.';
 describe('b-dropdown', () => {
   it('accessibility tests', async () => {
@@ -33,9 +33,12 @@ describe('b-dropdown', () => {
         </b-menu>
       </b-dropdown>
     `);
+    const trigger = element.shadowRoot?.querySelector('.dropdown__trigger') as HTMLSlotElement;
+
     expect(element.open).to.equal(false);
     // Mouse enter the dropdown trigger.
-    await sendMouse({ type: 'move', position: [20, 20] });
+
+    trigger.dispatchEvent(new MouseEvent('mouseenter', { bubbles: true }));
     await aTimeout(100);
     expect(element.open).to.equal(false);
     await aTimeout(100);
@@ -56,36 +59,18 @@ describe('b-dropdown', () => {
         </b-menu>
       </b-dropdown>
     `);
+    const trigger = element.shadowRoot?.querySelector('.dropdown__trigger') as HTMLSlotElement;
+    const content = element.shadowRoot?.querySelector('.dropdown__content') as HTMLElement;
+
     expect(element.open).to.equal(false);
     // Mouse enter the dropdown trigger.
-    await sendMouse({ type: 'move', position: [20, 20] });
+    trigger.dispatchEvent(new MouseEvent('mouseenter', { bubbles: true }));
     await aTimeout(100);
     expect(element.open).to.equal(true);
     // Mouse leave the dropdown content.
-    await sendMouse({ type: 'move', position: [0, 0] });
+    content.dispatchEvent(new MouseEvent('mouseleave', { bubbles: true }));
     await aTimeout(100);
     expect(element.open).to.equal(true);
-    await aTimeout(100);
-    expect(element.open).to.equal(false);
-  });
-
-  it('should not open the dropdown when mouse enter the trigger if disabled', async () => {
-    const element = await fixture<BDropdown>(html`
-      <b-dropdown disabled>
-        <b-button>Banana</b-button>
-        <b-menu slot="drop">
-          <b-menu-item>Option 1</b-menu-item>
-          <b-menu-item>Option 2</b-menu-item>
-          <b-menu-item>Option 3</b-menu-item>
-          <b-divider></b-divider>
-          <b-menu-item>Option A</b-menu-item>
-          <b-menu-item>Option B</b-menu-item>
-        </b-menu>
-      </b-dropdown>
-    `);
-    expect(element.open).to.equal(false);
-    // Mouse enter the dropdown trigger.
-    await sendMouse({ type: 'move', position: [20, 20] });
     await aTimeout(100);
     expect(element.open).to.equal(false);
   });
@@ -105,24 +90,27 @@ describe('b-dropdown', () => {
           </b-menu>
         </b-dropdown>
       `);
+      const trigger = element.shadowRoot?.querySelector('.dropdown__trigger') as HTMLSlotElement;
+      const content = element.shadowRoot?.querySelector('.dropdown__content') as HTMLElement;
+
       expect(element.open).to.equal(false);
       // Click the dropdown trigger.
-      await sendMouse({ type: 'click', position: [20, 20] });
+      trigger.dispatchEvent(new MouseEvent('click', { bubbles: true }));
       expect(element.open).to.equal(true);
 
       // Click the dropdown trigger again, the dropdown should be closed.
-      await sendMouse({ type: 'click', position: [20, 20] });
+      trigger.dispatchEvent(new MouseEvent('click', { bubbles: true }));
       expect(element.open).to.equal(false);
 
       // Make it open.
-      await sendMouse({ type: 'click', position: [20, 20] });
+      trigger.dispatchEvent(new MouseEvent('click', { bubbles: true }));
 
       // Click the content of the dropdown, the dropdown should not be closed.
-      await sendMouse({ type: 'click', position: [20, 60] });
+      content.dispatchEvent(new MouseEvent('click', { bubbles: true }));
       expect(element.open).to.equal(true);
 
       // Click the outside of the dropdown, the dropdown should be closed.
-      await sendMouse({ type: 'click', position: [0, 0] });
+      document.dispatchEvent(new MouseEvent('click', { bubbles: true }));
       expect(element.open).to.equal(false);
     });
 
@@ -140,17 +128,20 @@ describe('b-dropdown', () => {
           </b-menu>
         </b-dropdown>
       `);
+      const trigger = element.shadowRoot?.querySelector('.dropdown__trigger') as HTMLSlotElement;
+      const content = element.shadowRoot?.querySelector('.dropdown__content') as HTMLElement;
+
       expect(element.open).to.equal(false);
       // Mouse enter the dropdown trigger.
-      await sendMouse({ type: 'move', position: [20, 20] });
+      trigger.dispatchEvent(new MouseEvent('mouseenter', { bubbles: true }));
       expect(element.open).to.equal(false);
 
       // Click the dropdown trigger.
-      await sendMouse({ type: 'click', position: [20, 20] });
+      trigger.dispatchEvent(new MouseEvent('click', { bubbles: true }));
       expect(element.open).to.equal(true);
 
       // Should not close the dropdown when mouse leave the dropdown content.
-      await sendMouse({ type: 'move', position: [0, 0] });
+      content.dispatchEvent(new MouseEvent('mouseleave', { bubbles: true }));
       expect(element.open).to.equal(true);
     });
   });
@@ -170,10 +161,14 @@ describe('b-dropdown', () => {
           </b-menu>
         </b-dropdown>
       `);
+      const triggerElement = (
+        element.shadowRoot?.querySelector('.dropdown__trigger') as HTMLSlotElement
+      ).assignedElements()[0] as HTMLElement;
+
       expect(element.open).to.equal(false);
 
       // Focus the dropdown trigger.
-      (element.shadowRoot?.querySelector('.dropdown__trigger') as HTMLElement).focus();
+      triggerElement.focus();
       await sendKeys({ press: 'Enter' });
       expect(element.open).to.equal(true);
     });
@@ -192,10 +187,14 @@ describe('b-dropdown', () => {
           </b-menu>
         </b-dropdown>
       `);
+      const triggerElement = (
+        element.shadowRoot?.querySelector('.dropdown__trigger') as HTMLSlotElement
+      ).assignedElements()[0] as HTMLElement;
+
       expect(element.open).to.equal(false);
 
       // Focus the dropdown trigger.
-      (element.shadowRoot?.querySelector('.dropdown__trigger') as HTMLElement).focus();
+      triggerElement.focus();
       await sendKeys({ press: 'Enter' });
       expect(element.open).to.equal(true);
       // Press Escape.
@@ -217,10 +216,15 @@ describe('b-dropdown', () => {
           </b-menu>
         </b-dropdown>
       `);
+
+      const triggerElement = (
+        element.shadowRoot?.querySelector('.dropdown__trigger') as HTMLSlotElement
+      ).assignedElements()[0] as HTMLElement;
+
       expect(element.open).to.equal(false);
 
       // Focus the dropdown trigger.
-      (element.shadowRoot?.querySelector('.dropdown__trigger') as HTMLElement).focus();
+      triggerElement.focus();
       await sendKeys({ press: 'Space' });
       expect(element.open).to.equal(true);
       // Press Space.